Refactor Atom to use a single list of generic particles

+++ radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/Particle.java
@@ -1,9 +1,9 @@
 package com.radixdlt.client.core.atoms;
 
 public abstract class Particle {
-	private final int spin;
+	private final long spin;
 
-	Particle(int spin) {
+	Particle(long spin) {
 		this.spin = spin;
 	}
 }
